# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TempMail API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Domain {
  id: string;
  domain: string;
  isActive: boolean;
}

export interface MailboxInput {
  /** Full email address (username@domain) */
  address: string;
  password: string;
}

export interface Mailbox {
  id: string;
  address: string;
  token: string;
  createdAt: string;
}

export interface EmailAddress {
  address: string;
  name: string;
}

export interface MessageSummary {
  id: string;
  from: EmailAddress;
  subject: string;
  intro: string;
  seen: boolean;
  createdAt: string;
}

export interface Message {
  id: string;
  from: EmailAddress;
  to: EmailAddress[];
  subject: string;
  intro: string;
  /** @nullable */
  text?: string | null;
  /** @nullable */
  html?: string | null;
  seen: boolean;
  createdAt: string;
}