# lib/api-zod/src/generated/types/emailAddress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TempMail API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailAddress {
  address: string;
  name: string;
}