# lib/api-zod/src/generated/types/mailbox.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TempMail API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Mailbox {
  id: string;
  address: string;
  token: string;
  createdAt: string;
}