# lib/api-zod/src/generated/types/mailboxInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TempMail API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MailboxInput {
  /** Full email address (username@domain) */
  address: string;
  password: string;
}