# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TempMail API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailAddress } from './emailAddress';

export interface Message {
  id: string;
  from: EmailAddress;
  to: EmailAddress[];
  subject: string;
  intro: string;
  /** @nullable */
  text?: string | null;
  /** @nullable */
  html?: string | null;
  seen: boolean;
  createdAt: string;
}