# lib/api-zod/src/generated/types/messageSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TempMail API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailAddress } from './emailAddress';

export interface MessageSummary {
  id: string;
  from: EmailAddress;
  subject: string;
  intro: string;
  seen: boolean;
  createdAt: string;
}